3_2_2 Ready with Big Decimals

+++ polimorphism 3_2_2/src/Bill.java
@@ -1,18 +1,20 @@
 import ru.netology.taxtypes.TaxType;
 
+import java.math.BigDecimal;
+
 class Bill {
-    private double amount;
+    private BigDecimal amount;
     private TaxType taxType;
     private TaxService taxService;
 
-    public Bill(double amount, TaxType taxType, TaxService taxService) {
+    public Bill(BigDecimal amount, TaxType taxType, TaxService taxService) {
         this.amount = amount;
         this.taxType = taxType;
         this.taxService = taxService;
     }
 
     public void payTaxes() {
-        double taxAmount = taxType.calculateTaxFor(amount);
+        BigDecimal taxAmount = taxType.calculateTaxFor(amount);
         taxService.payOut(taxAmount);
     }
 }
